Now processing VEfile requirements
VEfile can now specify requirements in a very simple way. The
requirements will be installed. Also added some very high-level tests
to ensure that this works. Some things still to test are specifying git
requirements. However the unit tests for that are working so it's safe
to assume that is working correctly when used in VEfile. I promise I
will commit more rather than committing small portions after the fact.