Added more test tools
Added a temp_project context and a ContextUser object that allows you
to use contextmanagers without the 'with' clause. Also added a
FakeProject that is a subclass of Project but allows you to patch a
particular method of the project instance. This could be useful in a
more general sense but I don't want to get ahead of myself.